+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog of nens-auth-client
 0.13 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Fixed library with Django 4.*.
+
+- Fixed unittests with authlib 0.15.5.
 
 
 0.12 (2021-09-30)